worker/requirements.txt
playwright>=1.49.0
websockets>=13.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
httpx>=0.27.0
